// src/app/enums/team.enums.ts
export enum ETeamType {
  All = 'All',
  Staff = 'Staff',
  Internship = 'Internship',
  Collaborator = 'Collaborator'
}

export enum ETeamTypeId {
  All = 3,
  Staff = 0,
  Internship = 1,
  Collaborator = 2
}

export enum ERoleTypeId {
  Member = 0,
  PM = 1,
  Shadow = 2,
  Deactive = 3
}

export enum ERoleType {
  Member = 'Member',
  PM = 'PM',
  Shadow = 'Shadow',
  Deactive = 'Deactive'
}